Reorganization of the project layout
