+++ src/main/java/com/kodilla/ecommercee/controller/OrderController.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.controller;
 
 import com.kodilla.ecommercee.domain.OrderDto;
 import org.springframework.web.bind.annotation.*;
