+++ src/main/java/com/kodilla/ecommercee/controller/ProductController.java
@@ -1,6 +1,7 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.controller;
 
 
+import com.kodilla.ecommercee.domain.ProductDto;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
