+++ src/main/java/com/kodilla/ecommercee/domain/ProductDto.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.domain;
 
 public class ProductDto {
